minor
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@8698 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ blind/control-program.c
@@ -273,8 +273,8 @@ int main(int argc, char** args) {
 
         // compute scale range in arcseconds per pixel.
         
-		app_min = arcmin2arcsec(arcmin_width_min / imagew);
-        app_max = arcmin2arcsec(arcmin_width_max / imagew);
+		app_min = arcmin2arcsec(arcmin_width_min / (double)imagew);
+        app_max = arcmin2arcsec(arcmin_width_max / (double)imagew);
         solver->funits_lower = app_min;
         solver->funits_upper = app_max;
